4/3/20

+++ src/cs5004/animator/model/Rectangle.java
@@ -33,7 +33,7 @@ public class Rectangle extends AbstractShape {
    * @throws IllegalArgumentException if the String name is null
    */
 
-  public Rectangle(double x, double y, int r, int g, int b, double width, double height,
+  public Rectangle(int x, int y, int r, int g, int b, int width, int height,
           int startTime, int endTime, String name) throws IllegalArgumentException {
     super(x, y, r, g, b, width, height, startTime, endTime, name);
   }
@@ -49,8 +49,8 @@ public class Rectangle extends AbstractShape {
 
   @Override
   public String toString() {
-    return String.format("Name: %s\nType: Rectangle\nMin corner: (%.1f, %.1f), Width: %.1f, "
-                    + "Height: %.1f, Color: (%d, %d, %d)\nAppears at t=%d\nDisappears at t=%d\n\n",
+    return String.format("Name: %s\nType: Rectangle\nMin corner: (%d, %d), Width: %d, "
+                    + "Height: %d, Color: (%d, %d, %d)\nAppears at t=%d\nDisappears at t=%d\n\n",
             this.getName(), this.getX(), this.getY(), this.getWidth(), this.getHeight(),
             this.getR(), this.getG(), this.getB(), this.getStartTime(), this.getEndTime());
   }
